Extracted SelectAllFeature
